refactor: Structure in packages
Structure in packages to increase cohesiveness of related files and clean and modularize package tree for readability.
